move score-result-tool into dedicated page
